Final Lexical Analyzer Submission

+++ test.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    float a = 10;
+    int a = 10;
     int b = 20;
     char c = 'b'
     int sum = a + b;
